refactor: module-facade layout, drop client feature, panic + dedup hardening
- Module pattern X/{mod.rs facade, X.rs impl} across admin/db/executor/
  ingest/matcher/price; flat executor.rs -> executor/executor.rs
- Remove the `client` cargo feature (and root [features]); miden-client
  deps non-optional for simplicity
- P1/P2 fixes: replace crash-prone .expect()/panic sites (admin task,
  startup) with error logging + graceful cancel
- DB-first-insert dedup: insert_notes_batch reports first-insert ids;
  ingest gates order_tx on it; drop in-memory seen_notes HashSet
- Import hygiene: std::collections/time imports hoisted to file top
- solver.toml hygiene: add solver.toml.example, gitignore /solver.toml
- gitignore .DS_Store/.vscode noise
- New migrations: external_symbol on registered_tokens, onchain
  nullified status
- Integration tests moved under crates/solver/tests/

165/165 lib + 3/3 integration (release) green.

+++ crates/solver/src/matching/tests/test_experimental_v2.rs
@@ -195,7 +195,7 @@ fn dust_order_attack() {
     for order in engine.book.orders.values() {
         assert!(order.requested_filled() <= order.requested);
     }
-    assert_eq!(batch.remaining_orders, engine.book.active_order_count());
+    assert_eq!(batch.remaining_orders, engine.book.active_order_count() as u64);
 }
 
 /// Order splitting attack: instead of one 1000-unit order, place 1000 one-unit orders.
@@ -668,6 +668,6 @@ fn fuzz_adversarial_mix_500_trials() {
                 "trial {}: filled order {:?} has 0 fill", trial, oid
             );
         }
-        assert_eq!(batch.remaining_orders, engine.book.active_order_count());
+        assert_eq!(batch.remaining_orders, engine.book.active_order_count() as u64);
     }
 }
